feat: move recording recently visited into separate component
