+++ frontend/src/component/commandBar/CommandBar.tsx
@@ -14,7 +14,6 @@ import { useKeyboardShortcut } from 'hooks/useKeyboardShortcut';
 import { SEARCH_INPUT } from 'utils/testIds';
 import { useOnClickOutside } from 'hooks/useOnClickOutside';
 import { useOnBlur } from 'hooks/useOnBlur';
-import { useRecentlyVisited } from 'hooks/useRecentlyVisited';
 import {
     CommandResultGroup,
     type CommandResultGroupItem,
@@ -27,6 +26,7 @@ import { CommandFeatures } from './CommandFeatures';
 import { usePlausibleTracker } from 'hooks/usePlausibleTracker';
 import { CommandRecent } from './CommandRecent';
 import { CommandPages } from './CommandPages';
+import { RecentlyVisitedRecorder } from './RecentlyVisitedRecorder';
 
 export const CommandResultsPaper = styled(Paper)(({ theme }) => ({
     position: 'absolute',
@@ -110,7 +110,6 @@ export const CommandBar = () => {
     >([]);
     const [searchedFlagCount, setSearchedFlagCount] = useState(0);
     const [value, setValue] = useState<string>('');
-    const { lastVisited } = useRecentlyVisited();
     const { routes } = useRoutes();
     const allRoutes: Record<string, IPageRouteInfo> = {};
     for (const route of [
@@ -200,6 +199,7 @@ export const CommandBar = () => {
 
     return (
         <StyledContainer ref={searchContainerRef} active={showSuggestions}>
+            <RecentlyVisitedRecorder />
             <StyledSearch>
                 <SearchIcon
                     sx={{
@@ -266,10 +266,7 @@ export const CommandBar = () => {
                 elseShow={
                     showSuggestions && (
                         <CommandResultsPaper>
-                            <CommandRecent
-                                lastVisited={lastVisited}
-                                routes={allRoutes}
-                            />
+                            <CommandRecent routes={allRoutes} />
                             <CommandPageSuggestions routes={allRoutes} />
                         </CommandResultsPaper>
                     )
