+++ frontend/src/hooks/useRecentlyVisited.ts
@@ -1,6 +1,4 @@
 import { useCallback, useEffect, useState } from 'react';
-import { useLocation, useMatch } from 'react-router-dom';
-import { routes } from 'component/menu/routes';
 import { getLocalStorageItem, setLocalStorageItem } from '../utils/storage';
 import { basePath } from 'utils/formatPath';
 import { useCustomEvent } from './useCustomEvent';
@@ -20,8 +18,6 @@ const localStorageItems = (key: string): LastViewedPage[] => {
 
 export const useRecentlyVisited = () => {
     const key = `${basePath}:unleash-lastVisitedPages`;
-    const featureMatch = useMatch('/projects/:projectId/features/:featureId');
-    const projectMatch = useMatch('/projects/:projectId');
 
     const [lastVisited, setLastVisited] = useState<LastViewedPage[]>(
         localStorageItems(key),
@@ -33,26 +29,6 @@ export const useRecentlyVisited = () => {
             setLastVisited(localStorageItems(key));
         },
     );
-
-    const location = useLocation();
-
-    useEffect(() => {
-        if (!location.pathname) return;
-
-        const path = routes.find((r) => r.path === location.pathname);
-        if (path) {
-            setCappedLastVisited({ pathName: path.path });
-        } else if (featureMatch?.params.featureId) {
-            setCappedLastVisited({
-                featureId: featureMatch?.params.featureId,
-                projectId: featureMatch?.params.projectId,
-            });
-        } else if (projectMatch?.params.projectId) {
-            setCappedLastVisited({
-                projectId: projectMatch?.params.projectId,
-            });
-        }
-    }, [location, featureMatch, projectMatch]);
 
     useEffect(() => {
         if (lastVisited) {
